+++ src/C3.WinForms/State/LegacyGlobalState.vb
@@ -57,6 +57,7 @@ Module varGlobals
     Public brands As DataTable = catalogue.Tables("Brands")
     Public models As DataTable = catalogue.Tables("Models")
     Public tapes As DataTable = catalogue.Tables("Tapes")
+    Public ReadOnly brandService As New BrandService(New LegacyBrandRepository(Function() catalogue))
 
     Public deckCount As Integer
     Public brandCount As Integer